Worked on populators
Added Code to make Weapon Populator hopefully functional. Also added the text documents the populators will use to draw data from. In these text documents I included examples of how the data should be entered in to the text documents to add a new instance of that specific data type when populating. Additionally modified a few classes to include data that I decided would be useful.