Controller Work
Added Controller and moved related functions to it, also changed name of a few classes to avoid naming conflicts and setup for building the project into a JSON. Additionally made certain varaibles public so they could be modified and accessed by the controller. May in the future change all these back to protected or even private and make getter and setter methods in their respective classes.