Lot of Work
Worked on a lot of different aspects over the past few weeks, most importantly I got a lot of classes ready to work in the GUI. I got the first design fo the GUI finished and also implemented it so the first two pages work. Also for now the populators seem to be working and correctly reading in the information within the text files.